+++ packages/vtable/src/components/menu/dom/logic/MenuElement.ts
@@ -1,4 +1,4 @@
-import { EventHandler } from '../../../event/EventHandler';
+import { EventHandler } from '../../../../event/EventHandler';
 import type {
   DropDownMenuHighlightInfo,
   IDimensionInfo,
@@ -6,15 +6,15 @@ import type {
   MenuListItem,
   Placement,
   RectProps
-} from '../../../ts-types';
-import { createElement } from '../../../tools/dom';
+} from '../../../../ts-types';
+import { createElement } from '../../../../tools/dom';
 import { importStyle } from './MenuElementStyle';
-import { TABLE_EVENT_TYPE } from '../../../core/TABLE_EVENT_TYPE';
-import { cellInRange } from '../../../tools/helper';
-import { isValid } from '../../../tools/util';
-import type { PivotHeaderLayoutMap } from '../../../layout/pivot-header-layout';
-import { regUrl } from '../../../tools/global';
-import type { BaseTableAPI } from '../../../ts-types/base-table';
+import { TABLE_EVENT_TYPE } from '../../../../core/TABLE_EVENT_TYPE';
+import { cellInRange } from '../../../../tools/helper';
+import { isValid } from '../../../../tools/util';
+import type { PivotHeaderLayoutMap } from '../../../../layout/pivot-header-layout';
+import { regUrl } from '../../../../tools/global';
+import type { BaseTableAPI } from '../../../../ts-types/base-table';
 importStyle();
 
 const CLASSNAME = 'vtable__menu-element';
